Fixed THE LAST conversion warning found by clang.

+++ game/stage/character-stage-anim-num.cpp
@@ -98,7 +98,7 @@ namespace stage
                     const sf::Int8 ALPHA(static_cast<sf::Int8>(fadeCounter_));
                     if (fadeCounter_ < 254.0f)
                     {
-                        color_.a = static_cast<sf::Uint8>(255) - ALPHA;
+                        color_.a = static_cast<sf::Uint8>(255 - static_cast<int>(ALPHA));
                     }
                     else
                     {
